Make zoomAction a map util function
